Gets the mocha tests running in the browser.
Run `cake test:server`, and follow the instructions. The libs had to be
modified to work equally well in the browser and in NodeJS, and fpcalc needed
to be kept out completely (by returning unless a global `require` function is
available). That's OK, though, because it won't be run in the browser, but the
other modules will, and they're all passing.